// src/api/openapi/models/VehiclePhotoType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const VehiclePhotoType = {
    FRONT: 'Front',
    REAR: 'Rear',
    LEFT_SIDE: 'LeftSide',
    RIGHT_SIDE: 'RightSide',
    INTERIOR: 'Interior'
} as const;
export type VehiclePhotoType = typeof VehiclePhotoType[keyof typeof VehiclePhotoType];


export function instanceOfVehiclePhotoType(value: any): boolean {
    for (const key in VehiclePhotoType) {
        if (Object.prototype.hasOwnProperty.call(VehiclePhotoType, key)) {
            if (VehiclePhotoType[key as keyof typeof VehiclePhotoType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function VehiclePhotoTypeFromJSON(json: any): VehiclePhotoType {
    return VehiclePhotoTypeFromJSONTyped(json, false);
}

export function VehiclePhotoTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): VehiclePhotoType {
    return json as VehiclePhotoType;
}

export function VehiclePhotoTypeToJSON(value?: VehiclePhotoType | null): any {
    return value as any;
}

export function VehiclePhotoTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): VehiclePhotoType {
    return value as VehiclePhotoType;
}